+++ src/main/java/com/experisproject/experisproject/models/repositories/PersonRepository.java
@@ -3,5 +3,5 @@ package com.experisproject.experisproject.models.repositories;
 import com.experisproject.experisproject.models.entities.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person,Long> {
+public interface PersonRepository extends CrudRepository<Person,Integer> {
 }
